requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-rest-auth==2.2.5
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
